requirements.txt
paddlepaddle
paddleocr
pymupdf
pillow
fastapi
uvicorn
torch
transformers